Front, Roster and Apply are displaying correctly.
The Application form is working but may have internal bugs that haven't been sorted yet. This needs looking into.

The mobile design of all pages has suddenly stopped working. Media queries are not responding when the viewport becomes smaller but is working when you select the mobile version in chrome developer tools.